Ensure that numpy handled functions are loaded.
